requirements.txt
groq==0.13.0
python-dotenv==1.0.1
python_pptx==1.0.2
streamlit==1.39.0
